requirements.txt
flask~=3.0.3
flask_cors~=3.0.10
pydantic~=2.7.1
boto3~=1.34.113
botocore~=1.34.113
langchain~=0.1.20
python-dotenv~=1.0.1
langgraph~=0.0.48